refactor: mode ApiMetadataQueryService and fix arch tests https://gravitee.atlassian.net/browse/APIM-2583
