+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/architecture/OnionArchitectureTest.java
@@ -47,13 +47,16 @@ public class OnionArchitectureTest extends AbstractApimArchitectureTest {
             .domainModels(anyPackageThatContains(MODEL_PACKAGE))
             .domainServices(
                 "io.gravitee.apim.core.*." + CRUD_SERVICE_PACKAGE + "..",
-                "io.gravitee.apim.core.*." + DOMAIN_SERVICE_PACKAGE + ".."
+                "io.gravitee.apim.core.*." + DOMAIN_SERVICE_PACKAGE + "..",
+                "io.gravitee.apim.core.*." + QUERY_SERVICE_PACKAGE + ".."
             )
             .applicationServices(anyPackageThatContains(CORE_PACKAGE + ".*." + USECASE_PACKAGE))
             .adapter(
                 "persistence",
                 anyPackageThatContains(INFRA_PACKAGE + "." + CRUD_SERVICE_PACKAGE),
                 anyPackageThatContains(INFRA_PACKAGE + "." + DOMAIN_SERVICE_PACKAGE),
+                anyPackageThatContains(INFRA_PACKAGE + "." + QUERY_SERVICE_PACKAGE),
+                anyPackageThatContains(INFRA_PACKAGE + "." + NOTIFICATION_PACKAGE),
                 anyPackageThatContains(INFRA_PACKAGE + "." + ADAPTER_PACKAGE)
             )
             .adapter("spring", anyPackageThatContains(INFRA_PACKAGE + ".spring"))
